refactor: theme folder
