+++ packages/voodoo-ui/src/lib/components/Typography/utils/getColorToColorProps.ts
@@ -1,4 +1,4 @@
-import { VoodooUIBreakpointName } from '../../../theme/breakpoints';
+import { VoodooUIBreakpointName } from '../../../theme/voodooUI/breakpoints';
 import type {
   VoodooUIResponsiveRecordValue,
   VoodooUIResponsiveValue,
